introduce model_wrappers module
